Extract mixnet_connection_beacon
